# Conflicts:
#	resources/buildstock.rb
#	test/osm_files/MF_1story_UB_Furnace_AC1_FuelTankWH.osm
#	test/osm_files/MF_1story_UB_Furnace_AC1_FuelTankWH_MiddleUnit.osm
#	test/osm_files/MF_24units_2story_UB_3Beds_1pt5Baths_Denver_Furnace_CentralAC.osm
#	test/osm_files/MF_2story_UB_ASHP2_HPWH.osm
#	test/osm_files/MF_2story_UB_FuelBoiler_AC1_FuelTankWH.osm
#	test/osm_files/MF_2story_UB_Furnace_AC1_FuelTankWH.osm
#	test/osm_files/MF_2story_UB_Furnace_AC1_FuelTankWH_TopLevel.osm
#	test/osm_files/MF_HVACSizing_Load_8units_1story_UB.osm
#	test/osm_files/MF_Successful_EnergyPlus_Run_TMY_Appl_PV.osm
#	test/osm_files/SFA_10units_2story_SL_3Beds_1pt5Baths_Denver_Furnace_CentralAC.osm
#	test/osm_files/SFA_2story_FB_FuelBoiler_RoomAC_FuelTankWH.osm
#	test/osm_files/SFA_2story_UB_ASHP2_HPWH.osm
#	test/osm_files/SFA_2story_UB_FuelBoiler_RoomAC_FuelTankWH.osm
#	test/osm_files/SFA_2story_UB_Furnace_RoomAC_FuelTankWH.osm
#	test/osm_files/SFA_HVACSizing_Equip_Central_System_Boiler_Baseboards_Autosize.osm
#	test/osm_files/SFA_HVACSizing_Equip_Central_System_Fan_Coil_Autosize.osm
#	test/osm_files/SFA_HVACSizing_Equip_Central_System_PTAC_Autosize.osm
#	test/osm_files/SFA_HVACSizing_Load_4units_1story_FB_UA.osm
#	test/osm_files/SFD_1story_FB_UA_GRG_MSHP_FuelTanklessWH.osm
#	test/osm_files/SFD_1story_FB_UA_GRG_RoomAC_ElecBoiler_FuelTanklessWH.osm
#	test/osm_files/SFD_1story_UB_UA_ASHP2_HPWH.osm
#	test/osm_files/SFD_1story_UB_UA_GRG_ACV_FuelFurnace_HPWH.osm
#	test/osm_files/SFD_2story_CS_UA_AC2_FuelBoiler_FuelTankWH.osm
#	test/osm_files/SFD_2story_CS_UA_GRG_ASHPV_FuelTanklessWH.osm
#	test/osm_files/SFD_2story_FB_UA_GRG_AC1_ElecBaseboard_FuelTankWH.osm
#	test/osm_files/SFD_2story_FB_UA_GRG_AC1_UnitHeater_FuelTankWH.osm
#	test/osm_files/SFD_2story_FB_UA_GRG_GSHP_ElecTanklessWH.osm
#	test/osm_files/SFD_2story_PB_UA_ElecFurnace_ElecTankWH.osm
#	test/osm_files/SFD_2story_S_UA_GRG_ASHP1_FuelTanklessWH.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP1_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP1_AutosizeForMaxLoad_MinTemp.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP1_Autosize_MinTemp.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP1_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP2_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHP2_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHPV_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_ASHPV_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_BB_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_BB_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_Dehumidifier_Auto_Miami.osm
#	test/osm_files/SFD_HVACSizing_Equip_Dehumidifier_Fixed_Miami.osm
#	test/osm_files/SFD_HVACSizing_Equip_EF_AC2_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_EF_AC2_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_ElecBoiler_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_ElecBoiler_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_AC1_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_AC1_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_ACV_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_ACV_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_RAC_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GF_RAC_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GSHP_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GSHP_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_GasBoiler_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_GasBoiler_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHPDucted_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHP_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHP_AutosizeForMaxLoad.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHP_BB_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHP_BB_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_MSHP_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_UnitHeater_AC1_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_UnitHeater_AC1_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_UnitHeater_Autosize.osm
#	test/osm_files/SFD_HVACSizing_Equip_UnitHeater_Fixed.osm
#	test/osm_files/SFD_HVACSizing_Equip_UnitHeater_Fixed_wFan.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_PB_UA_Vented.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_PB_UA_Vented_ASHP_DuctsInPB.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_PB_UA_Vented_ASHP_DuctsInUA.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Unvented_InsRoof.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Unvented_NoOverhangs_NoIntShading_ERV.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Unvented_NoOverhangs_NoIntShading_HRV.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Unvented_NoOverhangs_NoIntShading_NoMechVent.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Unvented_NoOverhangs_NoIntShading_SupplyMechVent.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Vented.osm
#	test/osm_files/SFD_HVACSizing_Load_1story_S_UA_Vented_ExtFinDark.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FA.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FA_ASHP_DuctsInCS.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FA_ASHP_DuctsInGRG.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FA_ASHP_DuctsInLiv.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FA_Skylights.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_CS_GRG_FlatRoof_Skylights.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_FB_GRG_FA.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_FB_GRG_FA_ASHP_DuctsInFB.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_S_GRG_FA.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_UB_GRG_FA.osm
#	test/osm_files/SFD_HVACSizing_Load_2story_UB_GRG_FA_ASHP_DuctsInUB.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_AMY_PV.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_AMY_PV_TwoDays.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_TMY.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_TMY_Appl_PV.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_TMY_February.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_TMY_Outages.osm
#	test/osm_files/SFD_Successful_EnergyPlus_Run_TMY_Outages_PV.osm
#	test/test_samples_osw/base_results/results.csv

+++ Rakefile
@@ -193,17 +193,26 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
   end
   check_file_exists(lookup_file, nil)
 
+  lookup_csv_data = CSV.open(lookup_file, { col_sep: "\t" }).each.to_a
+
   # Perform various checks on each probability distribution file
   parameters_processed = []
-  tsvfiles = {}
   last_size = -1
 
   parameter_names = []
-  get_parameters_ordered_from_options_lookup_tsv(lookup_file).each do |parameter_name|
+  get_parameters_ordered_from_options_lookup_tsv(lookup_csv_data).each do |parameter_name|
     tsvpath = File.join(project_dir_name, housing_characteristics_dir, "#{parameter_name}.tsv")
     next if not File.exist?(tsvpath) # Not every parameter used by every project
 
     parameter_names << parameter_name
+  end
+
+  # Create all TsvFile objects for all parameters
+  tsvfiles = {}
+  parameter_names.each do |parameter_name|
+    tsvpath = File.join(project_dir_name, housing_characteristics_dir, "#{parameter_name}.tsv")
+    check_file_exists(tsvpath, nil)
+    tsvfiles[parameter_name] = TsvFile.new(tsvpath, nil)
   end
 
   while parameters_processed.size != parameter_names.size
@@ -214,15 +223,14 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
       err = "ERROR: Unable to process these parameters: #{unprocessed_parameters.join(', ')}."
       deps = []
       unprocessed_parameters.each do |p|
-        tsvpath = File.join(project_dir_name, housing_characteristics_dir, "#{p}.tsv")
-        tsvfile = TsvFile.new(tsvpath, nil)
-        tsvfile.dependency_cols.keys.each do |d|
+        tsvfiles[p].dependency_cols.keys.each do |d|
           next if deps.include?(d)
 
           deps << d
         end
       end
       undefined_deps = deps - unprocessed_parameters - parameters_processed
+
       # Check if undefined deps exist but are undefined simply because they're not in options_lookup.tsv
       undefined_deps_exist = true
       undefined_deps.each do |undefined_dep|
@@ -245,10 +253,7 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
       # Already processed? Skip
       next if parameters_processed.include?(parameter_name)
 
-      tsvpath = File.join(project_dir_name, housing_characteristics_dir, "#{parameter_name}.tsv")
-      check_file_exists(tsvpath, nil)
-      tsvfile = TsvFile.new(tsvpath, nil)
-      tsvfiles[parameter_name] = tsvfile
+      tsvfile = tsvfiles[parameter_name]
 
       # Dependencies not yet processed? Skip until a subsequent pass
       skip = false
@@ -296,13 +301,13 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
 
       # Check file format to be consistent with specified guidelines
       starting = Time.now
-      check_parameter_file_format(tsvpath, tsvfile.dependency_cols.length(), parameter_name)
+      check_parameter_file_format(tsvfile.full_path, tsvfile.dependency_cols.length(), parameter_name)
       ending = Time.now
       puts "  Checking file format: \t\t\t#{ending - starting} seconds\n"
 
       # Check for all options defined in options_lookup.tsv
       starting = Time.now
-      get_measure_args_from_option_names(lookup_file, tsvfile.option_cols.keys, parameter_name)
+      get_measure_args_from_option_names(lookup_csv_data, tsvfile.option_cols.keys, parameter_name, lookup_file)
       ending = Time.now
       puts "  Checking all options in options_lookup.tsv: \t#{ending - starting} seconds\n\n"
     end
@@ -313,7 +318,7 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
 
   # Test sampling
   r = RunSampling.new
-  output_file = r.run(project_dir_name, 1000, 'buildstock.csv', housing_characteristics_dir, lookup_file)
+  output_file = r.run(project_dir_name, 10000, "#{project_dir_name}.csv", housing_characteristics_dir, lookup_file)
 
   # Cache {parameter => options}
   parameters_options = {}
@@ -334,7 +339,7 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
   # Cache {parameter => {option => {measure => {arg => value}}}}
   parameters_options_measure_args = {}
   parameters_options.each do |parameter_name, option_names|
-    parameters_options_measure_args[parameter_name] = get_measure_args_from_option_names(lookup_file, option_names, parameter_name)
+    parameters_options_measure_args[parameter_name] = get_measure_args_from_option_names(lookup_csv_data, option_names, parameter_name, lookup_file)
   end
 
   # Check that measure arguments aren't getting overwritten
@@ -365,7 +370,11 @@ def integrity_check(project_dir_name, housing_characteristics_dir = 'housing_cha
   end
 
   if File.exist?(output_file)
-    File.delete(output_file) # Clean up
+    if project_dir_name == 'project_national'
+      FileUtils.mv(output_file, output_file.gsub(project_dir_name, 'buildstock'))
+    else
+      File.delete(output_file) # Clean up
+    end
   end
 
   # Unused TSVs?
@@ -394,20 +403,22 @@ def integrity_check_options_lookup_tsv(project_dir_name, housing_characteristics
   end
   check_file_exists(lookup_file, nil)
 
+  lookup_csv_data = CSV.open(lookup_file, { col_sep: "\t" }).each.to_a
+
   # Integrity checks for option_lookup.tsv
   measures = {}
   model = OpenStudio::Model::Model.new
 
   # Gather all options/arguments
-  parameter_names = get_parameters_ordered_from_options_lookup_tsv(lookup_file)
+  parameter_names = get_parameters_ordered_from_options_lookup_tsv(lookup_csv_data)
   parameter_names.each do |parameter_name|
     check_for_illegal_chars(parameter_name, 'parameter')
 
     tsvpath = File.join(project_dir_name, housing_characteristics_dir, "#{parameter_name}.tsv")
     next if not File.exist?(tsvpath) # Not every parameter used by every project
 
-    option_names = get_options_for_parameter_from_options_lookup_tsv(lookup_file, parameter_name)
-    options_measure_args = get_measure_args_from_option_names(lookup_file, option_names, parameter_name, nil)
+    option_names = get_options_for_parameter_from_options_lookup_tsv(lookup_csv_data, parameter_name)
+    options_measure_args = get_measure_args_from_option_names(lookup_csv_data, option_names, parameter_name, lookup_file)
     option_names.each do |option_name|
       check_for_illegal_chars(option_name, 'option')
